Revert "[Feat] HRV 모니터링 및 백그라운드 알림 (간이 테스트)"
