+++ Spha/Spha/Spha.entitlements
@@ -6,7 +6,5 @@
 	<true/>
 	<key>com.apple.developer.healthkit.access</key>
 	<array/>
-	<key>com.apple.developer.healthkit.background-delivery</key>
-	<true/>
 </dict>
 </plist>
