+++ Spha/Spha/SphaApp.swift
@@ -10,14 +10,17 @@ import SwiftUI
 @main
 struct SphaApp: App {
 
-    @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
     @StateObject var router: RouterManager = RouterManager()
+    let healthKitManager = HealthKitManager()
+    
+    init() {
+        configuire()
+    }
           
     var body: some Scene {
         WindowGroup {
             NavigationStack(path: $router.path){
-                // MainView()
-                HealthKitTestView()
+                MainView()
                     .navigationDestination(for: SphaView.self){ sphaView in
                         router.view(for: sphaView)
                     }
@@ -27,4 +30,7 @@ struct SphaApp: App {
         }
     }
 
+    func configuire() {
+        healthKitManager.requestAuthorization()
+    }
 }
